+++ frontend/lib/models/vote_model.dart
@@ -1,4 +1,5 @@
 class Vote {
+  int? id;
   String? subjectTitle;
   bool? repetition;
   bool? additional;
@@ -7,6 +8,7 @@ class Vote {
   List<int>? voteItemIds;
 
   Vote({
+    this.id,
     required this.subjectTitle,
     required this.repetition,
     required this.additional,
@@ -15,13 +17,27 @@ class Vote {
     this.voteItemIds,
   });
 
+  // 데이터를 가져올 때 사용
   factory Vote.fromJson(Map<String, dynamic> json) {
     return Vote(
+      id: json['id'],
       subjectTitle: json['subjectTitle'],
       repetition: json['repetition'],
       additional: json['additional'],
       announcementId: json['announcementId'],
-      endTime: json['endTime'],
+
+      // endTime이 List<int> 형식으로 전달되는 경우, 이를 DateTime으로 변환 후 String으로 변환
+      endTime: json['endTime']
+              is List<dynamic> // json['endTime']이 List<dynamic> 타입인지 확인
+          ? DateTime(
+              json['endTime'][0],
+              json['endTime'][1],
+              json['endTime'][2],
+              json['endTime'][3],
+              json['endTime'][4],
+            ).toIso8601String() // List<int>로 변환하고 DateTime 객체로 변환
+          : json['endTime'] as String?, // DateTime 객체를 ISO 8601 형식의 String으로 변환
+
       // 전달받은 json 맵에서 각 필드를 추출하여 Vote 객체 생성
       // voteItemIds는 List<dynamic> 형식으로 전달될 수 있기 때문에, 이를 List<int>로 변환하는 작업을 수행
       voteItemIds: (json['voteItemIds'] as List<dynamic>?)
@@ -30,6 +46,7 @@ class Vote {
     );
   }
 
+  // 요청 바디에 사용
   Map<String, dynamic> toJson() {
     return {
       'subjectTitle': subjectTitle,
@@ -43,5 +60,10 @@ class Vote {
       if (voteItemIds != null)
         'voteItemIds': voteItemIds!.map((item) => item.toInt()).toList(),
     };
+  }
+
+  @override
+  String toString() {
+    return '{id: $id, subjectTitle: $subjectTitle, repetition: $repetition, additional: $additional, announcementId: $announcementId, endTime: $endTime, voteItemIds: $voteItemIds}';
   }
 }
