+++ frontend/lib/providers/vote_provider.dart
@@ -6,6 +6,10 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'package:http/http.dart' as http;
 
 class VoteProvider with ChangeNotifier {
+  final List<Vote> _voteList = [];
+
+  List<Vote> get voteList => _voteList;
+
   // 엑세스 토큰 할당
   Future<String?> getToken() async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
@@ -33,12 +37,14 @@ class VoteProvider with ChangeNotifier {
         body: jsonEncode(vote.toJson()),
       );
 
-      final utf8Response = utf8.decode(response.bodyBytes);
-      final jsonResponse = jsonDecode(utf8Response);
-
-      final dataResponse = jsonResponse['data'];
+      print('Response status: ${response.statusCode}');
+      print('Response body: ${response.body}');
 
       if (response.statusCode == 201) {
+        final utf8Response = utf8.decode(response.bodyBytes);
+        final jsonResponse = jsonDecode(utf8Response);
+
+        final dataResponse = jsonResponse['data'];
         print('투표 생성 성공: $dataResponse');
       } else {
         print('투표 생성 실패: ${response.body}');
@@ -63,11 +69,24 @@ class VoteProvider with ChangeNotifier {
       },
     );
 
-    final utf8Response = utf8.decode(response.bodyBytes);
-    final jsonResponse = json.decode(utf8Response);
+    _voteList.clear();
 
     if (response.statusCode == 200) {
-      print('투표 조회 성공: ${response.body}');
+      final utf8Response = utf8.decode(response.bodyBytes);
+      final jsonResponse = json.decode(utf8Response);
+
+      final dataResponse = jsonResponse['data'];
+
+      if (dataResponse is List) {
+        // dataResponse가 List일 경우 각각의 item을 Vote 객체로 변환
+        for (var data in dataResponse) {
+          _voteList.add(Vote.fromJson(data));
+        }
+      } else if (dataResponse is Map<String, dynamic>) {
+        // dataResponse가 Map일 경우 직접 Vote 객체로 변환하여 추가
+        _voteList.add(Vote.fromJson(dataResponse));
+      }
+      print('투표 조회 성공: $_voteList');
     } else {
       print("투표 조회 실패: ${response.body}");
     }
