+++ frontend/lib/screens/boardDetail_screen.dart
@@ -1,12 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:frontend/all/providers/announcement_provider.dart';
+import 'package:frontend/models/vote_model.dart';
 import 'package:frontend/providers/vote_provider.dart';
+import 'package:frontend/widgets/vote_widget.dart';
 import 'package:frontend/services/login_services.dart';
 import 'package:provider/provider.dart';
 
 class BoardDetailPage extends StatefulWidget {
-  int? announcementId;
-  BoardDetailPage({this.announcementId, super.key});
+  final int? announcementId;
+  const BoardDetailPage({this.announcementId, super.key});
 
   @override
   State<BoardDetailPage> createState() => _BoardDetailPageState();
@@ -35,6 +37,7 @@ enum PopUpItem { popUpItem1, popUpItem2 } // 팝업 아이템
 
 class _BoardDetailPageState extends State<BoardDetailPage> {
   Map<String, dynamic> board = {};
+  Map<String, dynamic> voteMap = {}; // 투표 조회 변수
   String userRole = '';
   bool isLiked = false;
   int likeCount = 5;
@@ -58,11 +61,6 @@ class _BoardDetailPageState extends State<BoardDetailPage> {
         await Provider.of<AnnouncementProvider>(context, listen: false)
             .fetchOneBoard(widget.announcementId!);
       });
-
-      // WidgetsBinding.instance.addPostFrameCallback((_) async {
-      //   await Provider.of<VoteProvider>(context, listen: false)
-      //       .fetchVote(widget.announcementId!);
-      // });
     }
     _loadCredentials();
   }
@@ -141,27 +139,6 @@ class _BoardDetailPageState extends State<BoardDetailPage> {
                   ),
                 ],
               ),
-              // const Row(
-              //   children: [
-              //     Text(
-              //       '02/03',
-              //       style: TextStyle(
-              //         fontSize: 12,
-              //         fontWeight: FontWeight.normal,
-              //         color: Color(0xFFA89F9F),
-              //       ),
-              //     ),
-              //     SizedBox(width: 7),
-              //     Text(
-              //       '14:28',
-              //       style: TextStyle(
-              //         fontSize: 12,
-              //         fontWeight: FontWeight.normal,
-              //         color: Color(0xFFA89F9F),
-              //       ),
-              //     ),
-              //   ],
-              // ),
               const SizedBox(height: 20),
               Text(
                 board['announcementContent'],
@@ -178,7 +155,7 @@ class _BoardDetailPageState extends State<BoardDetailPage> {
                   width: 341,
                   height: 296,
                   child: Image.asset(
-                    'assets/images/classselect.png', // 파일 다운로드 api 때 적용해봐야 할 것 같음
+                    'assets/images/classselect.png',
                     fit: BoxFit.cover,
                   ),
                 ),
@@ -201,7 +178,7 @@ class _BoardDetailPageState extends State<BoardDetailPage> {
                       size: 20,
                     ),
                   ),
-                  const SizedBox(width: 2), // 아이콘과 텍스트 사이의 간격 조절
+                  const SizedBox(width: 2),
                   Text(
                     '$likeCount',
                     style: const TextStyle(
@@ -213,6 +190,26 @@ class _BoardDetailPageState extends State<BoardDetailPage> {
                 ],
               ),
               const SizedBox(height: 20),
+
+              // 투표 보기
+              if (board['votes'].isNotEmpty)
+                Theme(
+                  data: Theme.of(context)
+                      .copyWith(dividerColor: Colors.transparent),
+                  child: ExpansionTile(
+                    title: const Text('투표 보기'),
+                    children: [
+                      VoteWidget(
+                        votes: (board['votes'] as List<dynamic>)
+                            .map((vote) =>
+                                Vote.fromJson(vote as Map<String, dynamic>))
+                            .toList(),
+                      ),
+                    ],
+                  ),
+                ),
+
+              const SizedBox(height: 10),
               const Center(
                 child: Text(
                   '댓글을 작성할 수 없는 게시물입니다.',
